ai_pipeline/requirements.txt
sentence-transformers==3.3.1
transformers==4.47.1
torch>=2.1.0
faiss-cpu==1.9.0
scikit-learn==1.6.0
numpy>=1.24.0
